+++ SpriterDotNet/AnimationDataProvider/SnapshotAnimationDataProvider.cs
@@ -10,46 +10,73 @@ namespace SpriterDotNet.AnimationDataProvider
 {
     public class SnapshotAnimationDataProvider : DefaultAnimationDataProvider
     {
-        public static IDictionary<string, FrameData[]> GetData(SpriterEntity entity, int interval)
+        public static Dictionary<string, FrameData[]> CalculateData(SpriterEntity entity, int interval)
         {
-            IDictionary<string, FrameData[]> data = new Dictionary<string, FrameData[]>();
+            return Calculate<FrameData>(entity, interval, (d, a, t, i) => SpriterProcessor.UpdateFrameData(d, a, t));
+        }
+
+        public static Dictionary<string, FrameMetadata[]> CalculateMetadata(SpriterEntity entity, int interval)
+        {
+            return Calculate<FrameMetadata>(entity, interval, (d, a, t, i) => SpriterProcessor.UpdateFrameMetadata(d, a, t, i));
+        }
+
+        private static Dictionary<string, T[]> Calculate<T>(SpriterEntity entity, int interval, Action<T, SpriterAnimation, float, float> filler) where T : new()
+        {
+            Dictionary<string, T[]> results = new Dictionary<string, T[]>();
 
             foreach (SpriterAnimation anim in entity.Animations)
             {
                 int length = (int)Math.Ceiling(anim.Length / interval);
-                FrameData[] animData = new FrameData[length];
+                T[] animData = new T[length];
 
                 for (int i = 0; i < animData.Length; ++i)
                 {
                     float time = i * interval;
                     if (time > anim.Length) time = anim.Length;
 
-                    FrameData fd = new FrameData();
-                    SpriterProcessor.UpdateFrameData(fd, anim, time);
-                    animData[i] = fd;
+                    T data = new T();
+                    filler(data, anim, time, interval);
+                    animData[i] = data;
                 }
 
-                data[anim.Name] = animData;
+                results[anim.Name] = animData;
             }
-
-            return data;
+            return results;
         }
 
-        private readonly IDictionary<string, FrameData[]> data;
+        private readonly Dictionary<string, FrameData[]> data;
+        private readonly Dictionary<string, FrameMetadata[]> metaData;
 
-        public SnapshotAnimationDataProvider(IDictionary<string, FrameData[]> data)
+        public SnapshotAnimationDataProvider(Dictionary<string, FrameData[]> data, Dictionary<string, FrameMetadata[]> metaData)
         {
             this.data = data;
+            this.metaData = metaData;
+        }
+
+        public SnapshotAnimationDataProvider(SpriterEntity entity, int interval)
+        {
+            data = CalculateData(entity, interval);
+            metaData = CalculateMetadata(entity, interval);
         }
 
         public override FrameData GetFrameData(float time, float deltaTime, float factor, SpriterAnimation first, SpriterAnimation second = null)
         {
-            if (second != null) return base.GetFrameData(time, deltaTime, factor, first, second);
+            if (data == null || second != null) return base.GetFrameData(time, deltaTime, factor, first, second);
 
             FrameData[] animData = data[first.Name];
             int index = (int)(time / first.Length * animData.Length);
             if (index == animData.Length) index = animData.Length - 1;
             return animData[index];
+        }
+
+        public override FrameMetadata GetFrameMetadata(float time, float deltaTime, float factor, SpriterAnimation first, SpriterAnimation second = null)
+        {
+            if(metaData == null) base.GetFrameMetadata(time, deltaTime, factor, first, second);
+
+            FrameMetadata[] animMetadata = metaData[first.Name];
+            int index = (int)(time / first.Length * animMetadata.Length);
+            if (index == animMetadata.Length) index = animMetadata.Length - 1;
+            return animMetadata[index];
         }
     }
 }